Add initial language and dynamic semantics
Adds a core form of the Graceless language, currently without fields.
The language supports objects, qualified and unqualified requests, self
access, and a runtime store. Unqualified requests are supported in the
dynamic semantics with a substitution function which substitutes in a
receiver when it finds an unqualified request with a particular name. A
few example programs are also provided.