Ensure request substitution considers method names
Local request substitution looked at the parameters of methods to find
binding variables, but did not take into account the names of the
methods themselves. Now the subst metafunction ensures that, if an
object shadows a method, local requests inside that inner object refer
to the inner object's method.

Local requests to actual methods in object bodies are now bound when the
object is created, requiring that allocation be split into its
components at that point. This substitution goes straight to the
subst-method metafunction, to avoid the new restrictions in subst (that
is, to prevent the system from thinking methods are directly shadowing
themselves).

The examples have been reformatted and a new one has been added that
makes a local request to a surrounding method.