Replace store iterating functions list operations
The metafunctions that dealt with a store iterated through it like a
linked list. This is not necessary, as the standard list operations will
work on the store. This simplifies the fresh-location and lookup
metafunctions and splits set-field into a simple call to update-method,
which performs the method update on a single set of methods.