Make assignments optional instead of empty
Assignments to a variable declaration were either one of the two
assignments or empty. Now there is no empty assignment, and there are
two forms of vars, one with an assignment and one without. Either kind
of var is matched by the pattern (var x _ ...).

Store references are now required to be natural.