Add automatic test for errors, test all examples
The language is now automatically tested for errors with redex-test.
This generates example programs and ensures that the result is either a
value, an uninitialised error, or a stuck program. This catches cases
where the semantics are not completely defined. It caught that method
lookup could not handle non-unique method names in objects in the store,
which has now been prevented by a side-condition on both object creation
and requests.

All of the example programs are now explicitly tested to ensure they
produce an object with the right structure, or is stuck/produced an
error of the given term.