Add local variables to method bodies
Field declarations can now appear in method bodies as well, where they
are attached to the parameter object. To avoid exposing the parameter
object when a mutable variable is assigned to, assignments now return
the result of the assignment.

This change exposed a bug that local requests to assignment methods were
not being substituted correctly, which is now resolved.