Add extended language with object inheritance
The new language adds an inherits clause to objects, and adds the
expression in the clause as a hole in a context. It also extends the -->G
reduction relation to include copying the methods from the super object
into the sub-object. The resulting inheritance relation is concatenation:
modifying where self is bound at object creation time will change this
relation into either forwarding (self is fully bound when objects are
created) or delegation (self is bound in generated field methods when
objects are created). A number of examples and helper functions are also
included.

The result-equiv metafunction has also been extended to ensure it does not
care about the ordering of the method names it is given.