Remove irrelevant ignore matches in patterns
There were a number of pattern matches on methods that explicitly
ignored the different parts of the method, including ignoring all fields
with the (_ ...) pattern. This pattern will ignore any sequence of
anything, so in all of those cases the explicit patterns have been
replaced with just the sequence ignore.