Remove local variables, substitute parameters
Creating an object to hold potentially mutable variables when executing
a request was complicated and didn't add much to the model. Requests
arguments are now handled directly by regular substitution (to
unqualified requests with no arguments).

The application of Redex pattern syntax has been changed to be more
consistent, and rely less on non-terminal aliases for lists.