Implement delayed substitution in inherits
Local request substitution is no longer safe in the presence of
inherits, as the definition may be shadowed by a definition introduced
by the inherits clause. At runtime, inherits clauses are now qualified
by a collection of substitutions, which are run once the expression in
the inherits clause is evaluated. The definitions resulting from the
expression in the inherits clause are added to the substitution first,
to ensure those definitions take priority in the body of the object.