Test and fix the inheritance implementations
Two tests which operate differently depending on which inheritance
semantics is chosen have now been added. This ensures that they each
work as expected. Concatenation and delegation were over-eager in their
substitutions, and the base reduction relation substituted the wrong
reference for the inheriting object, all of which is now corrected. The
new substitution metafunction could do with some tidying up.