Tidy up substitution and object inheritance
The three different forms of substitution have been joined together into
a single operation by defining a non-terminal for substitutions that
ranges over the three different forms. As a result, applications of
substitution are a lot cleaner, and the operation is a lot easier to
parse.

The definitions are now more heavily commented, and many now make better
use of unquoting to use Racket list operations. These improvements have
been copied through into the inheritance implementations.