Add super reference, tidy up delayed substitution
The term super is now a keyword referring to the inherited object of the
current self object. As it is only substituted when the inherits clause
is reduced, no extra modifications are required to the separate
inheritance implementations.

Substitution over an object with an inherits clause now immediately
removes names which will be shadowed in the object, tidying up the
actual reduction. This change includes extending object-names to include
(but ignore) inherits clauses. The on the fly unquoting of shadow
removal operations have been moved into metafunctions because they each
occur more than once.