Add merged identity, elaborate super behaviour
Merged identity clones the object you are inheriting from, and then
concatenates the new methods into the original object. Only fresh
objects can be inherited from, which in enforced with a varying set of
contexts.

Tests for merged identity have revealed that super calls to fields
weren't really consistent in the existing implementations, so all of the
models except for concatenation now immediately forward to the super
object in inherited methods instead of copying the body. The
implementation of concatenation does not work completely correctly for
an inherited mutable field when the getter is overwritten. A test is
included which demonstrates that.