Add uniform identity
Uniform identity concatenates the body of an object literal in the
inherits clause in into the body of the inheriting object, with some
processing to removed overridden methods. A reasonable amount of
functionality has gone into dealing with the fact that there can now be
duplicate fields in an object, as the assignments from the fields are
still required to run in the specified order. Modifying the objects in
the store to explicitly store their fields, rather than relying on
method updates, will help alleviate this problem.