Update inheritance models to new language
This completes the previous changes to the core Graceless definition by
adding runtime fields and statements to the inheritance models. The
language definitions and reduction rules have been altered to bring them
in line with the text. Almost all of the rules correspond exactly to
their definition in the text, with some minor additions to deal with
making explicit some assumptions in the text.