Allow positional clauses to appear before methods
The body of an object under positional inheritance is now any body
element B, which is either a method M or statement S. The existing
behaviour is preserved by a call to split-body, which collects the
distinct methods and statements while preserving the statement ordering.
The inherits clauses now also have optional names, as with the simple
multiple inheritance. This means that positional is a super-set of all
of the single and simple multiple inheritance syntax, though not in
behaviour. The examples and table module have been updated accordingly.