Implement positional inheritance for object models
As with the basic multiple and method transformations, the object
inheritance systems are now supported as positional inheritance models.
This was not quite as straightforward as the others, because the syntax
of an object is different than from single inheritance, but involved
mostly copying the change across to forwarding and delegation.