Update base language to new Graceless definition
This change dramatically simplifies the process of building fields, at
the cost of user-facing syntax where getters and setters must be created
manually (the idea being that def and var are now just sugar for these
forms, except the sugar is not provided). Fields no longer exist at
runtime, and assignment performs a method update. Substitutions now only
contain a single form to substitute at a time (though subst can still
run on multiple substitutions for a single term). Types are also
included as well, though not the typing judgment.

The testing framework has been made easier to use: test-->>G is now
syntax, so the two arguments do not need to be quoted as terms, which
also means that the separate definitions can now use define-term
instead. The use of syntax also means that if the test fails it will
report the correct line number in the examples file, rather than always
giving the same position in the test framework file.