Add static judgments for the base language
Adds subtyping, signature selection, and typing judgments for Graceless.
This requires adding a store type, and a special self type form which is
not included in the paper rules (because the indexed self variables are
implicit there).

The new 'properties' module begins adding tests of standard soundness
properties based on the typing judgments, but currently it is difficult
for it to generate programs which are correctly typed.