Include type judgment in test cases
The tests now automatically test that the type of the outcome of any
reduction is a subtype of the type of the original term (though it does
not check every step along the way). The second argument to test-->>G is
now a type, rather than a list of method names, so the expected type of
the result is compared to the actual type of the result. As a result of
this testing, a number of mistakes in the existing rules have been
corrected.